New engine flag: "destroy-matches"
This flag tells the engine whether it is ok to destroy matches during
the construction of changes. This is useful when dealing with very large
amounts of multi-matches. For example, in the Sierpinski benchmark the
matches and the changes take up lots of memory. When setting this flag,
the engine can get rid of the matches during the construction of the
changes. Basically the freed memory of the matches can be used for the
change descriptions. I verified this positive effect in the Sierpinski
benchmark. By default, the flag is disabled though.

Change-Id: Ie916a54e6fcc598d9f72f941e0269f91bba8dd21
Signed-off-by: Christian krause <henshin.ck@gmail.com>