Replace singleton with regular list
The Singleton list does not support the retainAll operation which is
invoked a few lines after this change. In some situations retainAll can
be called on the singleton list causing an exception. This patch fixes
the problem.

I could not find a small reproducer of the error to add a test.

Change-Id: I1dcb4384bb61ec14c629d066064dc33942dd5bf7
Signed-off-by: Elie Richa <richa@adacore.com>