Added types to parameters as suggested by Christian Krause (dev mailing list on 01.11.14).
The new feature is named 'parameter kind' instead of 'parameter type' to
make it more distinguishable from object types. The following kinds are
supported: UNKNOWN (default), IN, OUT, INOUT, VAR

Please note that momentarily this is only of syntactic nature and has no
influence on the way that rules are interpreted. You can still freely
use parameters despite of their kind.

- Added enumeration to represent parameter kinds
- Added constraint that parameter names may not be equal to kind names
or aliases
- Regenerated model and edit projects
- Adapted non-generated model code
- Added icons to represent the different kinds in the tree-based editor
- Changed unit name parser to enable support in the graphical editor

Change-Id: I7a68a63221625cfbffa9bde7439ec1959b47631f
Signed-off-by: Stefan Schulz <schulzs@students.uni-marburg.de>