Bugfix in graphical editor: NACs/PACs in the node creation dialog
When specifying PAC or NAC nodes in the Node Creation dialog, there were
two bugs:

(1) Specifying nodes for the first NAC or PAC worked correctly. Nodes to
be included in additional PACs or NACs were added to the first PAC or
NAC instead. This was caused by lines 374-376 (now removed).

(2) Just entering the NAC name without the # prefix caused the editor to
handle the entered name as a multi-rule name, even if the "multi" option
was deselected. This was caused by lines 369-371 (now fixed).

Change-Id: Ic83a88e283f49b0de1a4b96430561b1046bc2b3e