Fix behavior when creating a new edge between two <<preserve>> nodes
The fix enables the specification of <<create>> and <<forbid>> edges
between two <<preserve>> nodes at the same time (see [1]). To this end,
it changes the default action of a newly added edge to <<create>>, if
the new edge is added between two <<preserve>> nodes with a <<forbid>>
edge between them.

[1] https://dev.eclipse.org/mhonarc/lists/henshin-user/msg00056.html

Change-Id: Ia5a1a74d496228f91cda15b6c06719dcf05db9a8