Workaround for bug 521619.
When changing the subunit, the editor first checks, whether it
needs to delete the graphical representation of the subunit.
If the subunit was changed, e.g. by moving it, this check fails
and a duplicate of the part is created.

As a work-around for this issue, the editor determines whether the part
should be deleted by checking for signature identity instead of
object identity.
In this context, two units are deemed identical if their names,
parameter names, and parameter types match.

Change-Id: Iad5aecf541848882aa82229239b8a197a1f05744
Signed-off-by: Stefan Schulz <schulzs@mathematik.uni-marburg.de>