[Build] Add quality-gates and log tests in Jenkins pipeline
Enable the creation of required compiler logs and enable the display of
warnings in them so that they can be considered in the quality-gates.

Also don't fail the entire build on test-failures but let Jenkins record
all tests and mark the build as unstable (instead of just failed) in
case tests fail.

Besides the two states 'success' and 'failed' the Jenkins build also has
the third state 'unstable' that is entered if the actual build succeeds
and only tests have failed or the qulity-gate was missed.
This means that the build only reaches the failed state on more severe
problems like compile-errors, which makes the quick assessment of
problems easier.

Part of
- https://github.com/eclipse-henshin/henshin/issues/5