Fix file encoding errors
Multiple files seems to have a wrong encoding, according to the Jenkins
fingerprint computer.
For some files it's sufficient to just touch them
and in some cases wrongly encoded characters (which are just shown with
the Unicode replacement character) are replaced by the (assumed) correct
character (mostly German Sonderzeichen).

In a few cases, fixing the encoding revealed other problems in the
corresponding files that are fixed too.